♻️ [Administrateur] Migrated entity to attributes
