--- config/doctrine/Administrateur.orm.yml
@@ -1,31 +1,0 @@
-App\Entity\Administrateur:
-    type: entity
-    table: "administrateur"
-    repositoryClass: App\Repository\AdministrateurRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        createdAt:
-            type: datetime
-            column: createdAt
-            gedmo:
-                timestampable:
-                    on: create
-        updatedAt:
-            type: datetime
-            column: updatedAt
-            gedmo:
-                timestampable:
-                    on: update
-    oneToOne:
-        user:
-            targetEntity: App\Entity\User
-            cascade: [persist,remove]
-            inversedBy: subjectAdministrateur
-            joinColumn:
-                nullable: false
-    lifecycleCallbacks: {  }
