+++ src/Entity/User.php
@@ -14,6 +14,7 @@ use App\Api\State\UserPasswordProcessor;
 use App\Api\State\UserStateProcessor;
 use App\Controller\Api\MeController;
 use App\Entity\Attributes\AccessToken;
+use App\Entity\Attributes\Administrateur;
 use App\Entity\Attributes\Adresse;
 use App\Entity\Attributes\Association;
 use App\Entity\Attributes\Centre;
